refactor: move code to plugins module, infer better names for modules
